pyproject.toml
[project]
name = "langchain-agents"
version = "0.1.0"
dependencies = [
    "langchain",
    "langchain-openai",
    "python-dotenv"
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"